GDB-10391 - add debounce utility class
