+++ src/js/angular/controllers.js
@@ -23,6 +23,7 @@ import './guides/directives';
 import {GUIDE_PAUSE} from './guides/tour-lib-services/shepherd.service';
 import 'angular-pageslide-directive/dist/angular-pageslide-directive';
 import 'angularjs-slider/dist/rzslider.min';
+import {DebounceUtils} from "./utils/debounce-utils";
 
 angular
     .module('graphdb.workbench.se.controllers', [
@@ -201,7 +202,7 @@ function mainCtrl($scope, $menuItems, $jwtAuth, $http, toastr, $location, $repos
         $scope.initTutorial();
     });
 
-    $scope.checkMenu = debounce(function() {
+    $scope.checkMenu = DebounceUtils.debounce(function() {
         return $('.main-menu').hasClass('collapsed');
     }, 250);
 
@@ -717,40 +718,6 @@ function mainCtrl($scope, $menuItems, $jwtAuth, $http, toastr, $location, $repos
         } else {
             collapsedMenuLogicOnInit();
         }
-    }
-
-    /**
-     * Creates a debounced function that delays invoking the provided function until after a specified wait time has
-     * elapsed since the last time the debounced function was invoked. Optionally, the function can be invoked immediately
-     * and then wait for the specified time before allowing the next invocation.
-     * @param {Function} func - The function to debounce.
-     * @param {number} wait - The number of milliseconds to delay.
-     * @param {boolean} [callImmediately=false] - If true, the function will be invoked immediately, then wait for the specified time
-     *                                           before allowing the next invocation.
-     * @return {Function} - Returns the new debounced function.
-     */
-    function debounce(func, wait, callImmediately) {
-        let timeout;
-
-        return function(...args) {
-            const later = () => {
-                timeout = null;
-                if (!callImmediately) {
-                    func.apply(this, args);
-                }
-            };
-            const callNow = callImmediately && !timeout;
-
-            if (timeout) {
-                clearTimeout(timeout);
-            }
-
-            timeout = setTimeout(later, wait);
-
-            if (callNow) {
-                func.apply(this, args);
-            }
-        };
     }
 
     $scope.slideToPage = function (index) {
